section3

+++ src/main/java/com/example/entity/BookDTO.java
@@ -9,10 +9,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookDTO {
-    private Long num;
+    private int num;
     private String title;
     private int price;
     private String author;
     private int page;
-
 }
